%post, %preun, %check scripts to exit 0 by default.
Explanation: The normal behaviour of shell is to exit with the exit code 
of the last-run command.  Sometimes this means the last 'if' condition, 
or a '&&' conjunction can unintentionally abort an install. This change
effectively adds an 'exit 0' to the end of most scripts. To raise a failure,
scripts must now explicitly call 'exit 1'. This will terminate the 
installation process on most platforms.