Add option to strip ELF binaries
Unstripped binaries can be separated out into a tarball
with the --save-unstripped option. They are can be drop-in replacements
for binaries in the packages.

One caveat: the binaries in DESTDIR are stripped in-place, so
if the --save-unstripped option is used and polypkg is re-run on the
same DESTDIR, it won't find any unstripped binaries. This can be worked
around be re-running "make install" from the build dir. It could
probably be solved by hardlink farming a copy of the destdir, making a
copy of each unstripped ELF object before stripping it.