[deb] Strip invalid characters from version
Package versions are now compliant with Debian policy section 5.6.12 and
will not cause dpkg 1.15.8 and later to fail to create the package.

dpkg 1.15.8 and later are stricter about version numbers: refusing to
build packages with invalid version numbers and warning about existing
packages with invalid version numbers.

If the user specifies the version explicitly with the pp_deb_version
variable, it will not be sanitized. This could be changed, but at the
moment I figure it makes sense to let the user provide whatever version
string they want if they're overriding it for the deb target.