pp.back.systemd.svc: fix false detection of init.d/service status
Normal init.d status script returns with non zero if service is not running.
However, on sles12 at least, systemd status gets called instead, which
prints the info on the screen about service is not running and returns
with success (0).

This results in the service getting "restarted" even though it was not
running before.