pp.back.systemd: check status and restart with systemctl
If the service is not running according to its sysv script (eg.
systemd service is not using a pidfile the sysv script needs), or
its sysv script is not present any more (the next upgrade), we also
check how systemd thinks, and stop with that.

We also try to restart the service in case something went wrong with it
(if it is in a failed state).